rename random => rng
