+++ openpathsampling/rng.py
